rapidyaml: Add parsing of ys to integer-based events
Adds a parser producing integer events, and integers signifying strings
by indexing into the parsed YS string.

Each event is provided as an integer bitmask. The event bits are
defined in both the C++ side (ysparse_evt_handler.hpp) and the Java
side (org.rapidyaml.Evt), and they need to stay consistent on both
ends. When a string is associated with an event, it is provided as an
integer offset and length.

For example, the YAML `say: 2 + 2` produces the following sequence of
integers:

```c++
BSTR,
BDOC,
VAL|BMAP|BLCK,
KEY|SCLR|PLAI, 0, 3, // "say"
VAL|SCLR|PLAI, 5, 5, // "2 + 2"
EMAP,
EDOC,
ESTR,
```

Note that the scalar events, ie "say" and "2 + 2", are followed each
by two extra integers encoding the offset and length of the scalar's
string. These two extra integers are present whenever the event has
any of the bits SCLR, ALIA, ANCH or TAG. For ease of use, there is a
bitmask HAS_STR, which enables quick testing by a simple `flags & HAS_STR`.

Also, where a string requires filtering, the parser filters it
in-place in the input string, and returns the extra integers
pertaining to the resulting filtered string.

The existing EDN-producing parser was not removed, and the
EVT-producing parser was added both on C++ and on the Java JNI bridge.

Tests were enlarged to cover the new event parsing, both for C++ and
Java.

To benefit from this approach, the YS side must implement a mechanism
to convert the int event sequence into its internal data-structure,
using the symbols in the class org.rapidyaml.Evt.

Other changes:

- rapidyaml->ysparse: start renaming the C++ library providing the
  JNI. ysparse is a more accurate description of what the library does.
- Directly use only the required rapidyaml source files instead of
  amalgamating into a single header.
- Makefile:
  - Add target to generate the JNI header
  - Improve target dependencies